zap/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
